🚚 Consolidate auth into features/auth and lib wiring
Move auth domain logic (blocked/temp email policy, anonymous user
merging) into features/auth, fold features/password into it, move the
Better-Auth plugin to lib/auth-plugins and the iron-session token
helpers to lib. Deletes src/auth and src/utils/session (RAL-1289).
