+++ apps/web/src/app/[locale]/(auth)/reset-password/components/reset-password-form.tsx
@@ -14,8 +14,8 @@ import { useRouter, useSearchParams } from "next/navigation";
 import { useForm } from "react-hook-form";
 import * as z from "zod";
 import { PasswordInput } from "@/components/password-input";
-import { PasswordStrengthMeter } from "@/features/password/components/password-strength-meter";
-import { usePasswordValidationSchema } from "@/features/password/schema";
+import { PasswordStrengthMeter } from "@/features/auth/components/password-strength-meter";
+import { usePasswordValidationSchema } from "@/features/auth/schema";
 import { Trans } from "@/i18n/client";
 import { authClient } from "@/lib/auth-client";
 import { validateRedirectUrl } from "@/utils/redirect";
