+++ apps/web/src/app/[locale]/(space)/settings/security/components/change-password-form.tsx
@@ -13,8 +13,8 @@ import { toast } from "@rallly/ui/sonner";
 import { useForm } from "react-hook-form";
 import * as z from "zod";
 import { PasswordInput } from "@/components/password-input";
-import { PasswordStrengthMeter } from "@/features/password/components/password-strength-meter";
-import { usePasswordValidationSchema } from "@/features/password/schema";
+import { PasswordStrengthMeter } from "@/features/auth/components/password-strength-meter";
+import { usePasswordValidationSchema } from "@/features/auth/schema";
 import { Trans, useTranslation } from "@/i18n/client";
 import { authClient } from "@/lib/auth-client";
