+++ apps/web/src/features/auth/components/password-strength-meter.tsx
