+++ apps/web/src/lib/auth.ts
@@ -23,17 +23,16 @@ import { headers } from "next/headers";
 import { redirect } from "next/navigation";
 import { after } from "next/server";
 import { cache } from "react";
-import { isEmailBlocked } from "@/auth/helpers/is-email-blocked";
-import { linkAnonymousUser } from "@/auth/helpers/merge-user";
-import { isTemporaryEmail } from "@/auth/helpers/temp-email-domains";
-import { hostOnlyCookieCleanup } from "@/auth/plugins/host-only-cookie-cleanup";
 import { getInstanceBranding } from "@/emails/branding";
 import { env } from "@/env";
 import { posthog } from "@/features/analytics/posthog";
+import { linkAnonymousUser } from "@/features/auth/mutations";
+import { isEmailBlocked, isTemporaryEmail } from "@/features/auth/utils";
 import { createSpace } from "@/features/space/mutations";
 import type { UserDTO } from "@/features/user/schema";
 import { getTranslation } from "@/i18n/server";
 import { getLocale } from "@/i18n/server/get-locale";
+import { hostOnlyCookieCleanup } from "@/lib/auth-plugins/host-only-cookie-cleanup";
 import { redis } from "@/lib/kv";
 import { getValueByPath } from "@/utils/get-value-by-path";
